+++ python/qibuild/actions/clean.py
@@ -17,6 +17,8 @@ import os
 
 import qisys
 from qisys import ui
+from qisys import interact
+
 import qibuild
 import qibuild.cmdparse
 
@@ -25,19 +27,11 @@ def configure_parser(parser):
     qibuild.parsers.toc_parser(parser)
     qibuild.parsers.build_parser(parser)
     qibuild.parsers.project_parser(parser)
-    parser.add_argument("-z", dest="all_toolchains", action="store_true",
-                                                help="erase for all toolchains")
+    parser.add_argument("-z", dest="remove_known_configs", action="store_true",
+                        help="remove all build directories that match a known configuration")
+    parser.add_argument("-x", dest="remove_unknown_configs", action="store_true",
+                        help="remove build directories that do not match any known configuration")
     parser.add_argument("--force", "-f", dest="force", action="store_true", help="force the clean")
-
-def only_existing_dirs(dirs):
-    """Return a list of existing directories from a list of dirs."""
-    existing_dirs = [x for x in dirs if os.path.isdir(x)]
-    return existing_dirs
-
-def get_build_dirs(projects):
-    """Returns a list of build directories."""
-    bdirs = [project.build_directory for project in projects]
-    return bdirs
 
 @ui.timer("qibuild clean")
 def do(args):
@@ -45,31 +39,69 @@ def do(args):
     toc = qibuild.toc.toc_open(args.worktree, args)
     (_, projects) = qibuild.cmdparse.deps_from_args(toc, args)
 
-    if args.all_toolchains:
-        bdirs = list()
-        for project in projects:
-            dirs = os.listdir(project.path)
-            bdirs.extend([os.path.join(project.path, x) for x in dirs if
-                os.path.isdir(os.path.join(project.path, x)) and
-                                      qibuild.project.is_build_dir(project, x)])
+    if args.remove_known_configs and args.remove_unknown_configs:
+        clean_selection= "all_configs"
+    elif args.remove_known_configs and not args.remove_unknown_configs:
+        clean_selection = "known_configs"
+    elif not args.remove_known_configs and args.remove_unknown_configs:
+        clean_selection = "unknown_configs"
     else:
-        bdirs = get_build_dirs(projects)
+        clean_selection = "given_config"
 
-    bdirs = only_existing_dirs(bdirs)
-    bdir_count = len(bdirs)
+    bdirs = {'known_configs': list(), 'unknown_configs': list()}
+    all_configs = clean_selection != "given_config"
+    for project in projects:
+        bdirs_ = qibuild.project.get_build_dirs(project, all_configs)
+        for cat in bdirs_.keys():
+            bdirs[cat].extend(bdirs_[cat])
 
-    if bdir_count == 0:
-        ui.info(ui.green, "No directory to clean")
-    elif not args.force:
-        ui.info(ui.green, "Build directories that will be removed", ui.reset, ui.bold, "(use -f to apply):")
+    if clean_selection in ["given_config", "all_configs", "known_configs"]:
+        bdir_count = len(bdirs['known_configs'])
+        if bdir_count == 0:
+            ui.info(ui.green, "No build directory to clean")
+        elif not args.force:
+            ui.info(ui.green, "Build directories that will be removed",
+                ui.reset, ui.bold, "(use -f to apply):")
+        for i, bdir in enumerate(bdirs['known_configs'], start=1):
+            to_print = [ui.green, "*", ui.reset, "(%i/%i)" % (i, bdir_count)]
+            if args.force:
+                to_print.extend([ui.green, "Cleaning", ui.reset, bdir])
 
-    for i, bdir in enumerate(bdirs, start=1):
-        to_print = [ui.green, "*", ui.reset, "(%i/%i)" % (i, bdir_count)]
-        if args.force:
-            to_print.extend([ui.green, "Cleaning", ui.reset, bdir])
+                # delete the build directory
+                qisys.sh.rm(bdir)
+            else:
+                to_print.extend([ui.reset, bdir])
+            ui.info(*to_print)
 
-            # delete the build directory
-            qisys.sh.rm(bdir)
-        else:
-            to_print.extend([ui.reset, bdir])
-        ui.info(*to_print)
+    if clean_selection in ["all_configs", "unknown_configs"]:
+        bdir_count = len(bdirs['unknown_configs'])
+        if bdir_count == 0:
+            ui.info(ui.green, "No build directory matching unknown configuration to clean")
+        elif not args.force:
+            ui.info(ui.green, "Build directories matching unknown configuration that may be removed",
+                ui.reset, ui.bold, "(interactive mode, use -f to apply):")
+        # remove uncertain build directories, by configuration name, so sort them
+        sorted_bdirs = {}
+        for bdir in bdirs['unknown_configs']:
+            # all build directory names should be prefixed with "build-", so strip it
+            config_name = os.path.basename(bdir)[6:]
+            if not config_name in sorted_bdirs:
+                sorted_bdirs[config_name] = []
+            sorted_bdirs[config_name].append(bdir)
+        for c, sbdirs in sorted_bdirs.items():
+            question = "Remove build directories matching the '%s' configuration?" % c
+            answer = interact.ask_yes_no(question, default=False)
+            if not answer:
+                continue
+            bdir_count = len(sbdirs)
+            for i, bdir in enumerate(sbdirs, start=1):
+                to_print = [ui.green, "*", ui.reset, "(%i/%i)" % (i, bdir_count)]
+                if args.force:
+                    to_print.extend([ui.green, "Cleaning", ui.reset, bdir])
+
+                    # delete the build directory
+                    qisys.sh.rm(bdir)
+                else:
+                    to_print.extend([ui.reset, bdir])
+                ui.info(*to_print)
+